+++ pkg/server/deploy/deploy.go
@@ -35,7 +35,7 @@ type Operations interface {
 type K8sOperations interface {
 	CreateOrUpdateDeploy(deploySpec *spec.Deploy) error
 	CreateOrUpdateCronJob(cronJobSpec *spec.CronJob) error
-	ExposeDeploy(namespace, name, vHost, svcType string, w io.Writer) error
+	ExposeDeploy(namespace, name, vHost, svcType, portName string, w io.Writer) error
 	ReplicaSetListByLabel(namespace, label, value string) ([]*ReplicaSetListItem, error)
 	DeployRollbackToRevision(namespace, name, revision string) error
 	CreateOrUpdateConfigMap(namespace, name string, data map[string]string) error
@@ -229,7 +229,7 @@ func (ops *DeployOperations) exposeApp(a *app.App, w io.Writer) error {
 		return nil
 	}
 	svcType := ops.serviceType(a)
-	if err := ops.k8s.ExposeDeploy(a.Name, a.Name, a.VirtualHost, svcType, w); err != nil {
+	if err := ops.k8s.ExposeDeploy(a.Name, a.Name, a.VirtualHost, svcType, a.Protocol, w); err != nil {
 		return err
 	}
 	return nil // already exposed