+++ pkg/server/spec/service_test.go
@@ -5,8 +5,9 @@ import "testing"
 func TestNewDefaultService(t *testing.T) {
 	expectedAppName := "test"
 	expectedServiceType := "LoadBalancer"
+	expectedProtocol := "protocol"
 
-	ss := NewDefaultService(expectedAppName, expectedServiceType)
+	ss := NewDefaultService(expectedAppName, expectedServiceType, expectedProtocol)
 	if ss.Namespace != expectedAppName {
 		t.Errorf("expected %s, got %s", expectedAppName, ss.Namespace)
 	}
@@ -18,6 +19,9 @@ func TestNewDefaultService(t *testing.T) {
 	}
 	if ss.Ports[0].TargetPort != DefaultPort {
 		t.Errorf("expected %d, got %d", DefaultPort, ss.Ports[0].TargetPort)
+	}
+	if ss.Ports[0].Name != expectedProtocol {
+		t.Errorf("expected %s, got %s", expectedProtocol, ss.Ports[0].Name)
 	}
 	if name := ss.Labels["run"]; name != expectedAppName {
 		t.Errorf("expected label run with value %s, got %s", expectedAppName, name)